ai110-module3show-musicrecommendersimulation-starter/requirements.txt
anthropic>=0.40.0
pandas
pytest
python-dotenv
streamlit
tabulate
